+++ nengo/tests/test_config.py
@@ -1,4 +1,5 @@
 import pickle
+import platform
 from fnmatch import fnmatch
 
 import pytest
@@ -159,6 +160,10 @@ def test_configstack():
     assert inhibit.transform.init == -1
 
 
+@pytest.mark.skipif(
+    platform.system() not in ["Linux", "Windows"],
+    reason="TODO: Not working on Mac OS X",
+)
 def test_config_property():
     """Test that config can't be easily modified."""
     with nengo.Network() as net:
@@ -302,6 +307,10 @@ def test_config_repr():
     )
 
 
+@pytest.mark.skipif(
+    platform.system() not in ["Linux", "Windows"],
+    reason="TODO: Not working on Mac OS X",
+)
 def test_config_exit_errors():
     """Tests ConfigErrors in `Config.__exit__`"""
     model = nengo.Network()